requirements.txt
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==3.3.2
idna==2.8
isodate==0.6.0
msrest==0.6.10
oauthlib==3.1.0
psutil==5.6.6
pycparser==2.19
pyserial==3.4
python-can==3.3.4
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
